refactor: codebase and minimal changes
