+++ Cargo.toml
@@ -1,9 +1,9 @@
 [package]
 name = "trimsec"
-description = "Strategic (& fast) content consumption planner."
+description = "Calculate time saved on using speed multipliers."
 authors = ["HitBlast <hitblastlive@gmail.com>"]
-version = "0.3.1"
-edition = "2021"
+version = "1.0.0"
+edition = "2024"
 repository = "https://github.com/hitblast/trimsec"
 homepage = "https://github.com/hitblast/trimsec"
 license = "MIT"
