+++ README.md
@@ -5,46 +5,26 @@
 [![Unit Tests](https://github.com/hitblast/trimsec/actions/workflows/tests.yml/badge.svg)](https://github.com/hitblast/trimsec/actions/workflows/tests.yml)
 [![Release Builds](https://github.com/hitblast/trimsec/actions/workflows/release.yml/badge.svg)](https://github.com/hitblast/trimsec/actions/workflows/release.yml)
 
-Strategic (& fast) content consumption planner.<br>
-
-> [!NOTE]
-> This project is still under active development. New ideas of improvement are always welcome!
+Calculate time saved on using speed multipliers.
 
 ## Table of Contents
 
 - [Overview](#overview)
 - [Usage](#usage)
 - [Installation](#installation)
-  - [cargo](#cargo) [rust]
-  - [mise](#mise) [rust]
-- [Manual installation](#manual-installation)
+  - [cargo](#cargo)
+  - [mise](#mise)
+- [Manual Installation](#manual-installation)
 - [Contributors](#contributors)
-- [Changelog](#changelog)
 - [License](#license)
 
 ## Overview
 
-[trimsec](https://crates.io/crates/trimsec) provides an easy way to accurately
-estimate how much time you save on watching videos when using speed multipliers.
-This is essentially helpful when watching recorded classes or documentaries
-online, in case you do not have enough time to spare based on your situation.
-
-Since I am also a student and both coordinating between my routine, as well as
-manually doing the math required to estimate my overall hours for study wasn't
-an option I'd consider - I chose to program a tool that would automate the
-process for me.
-
-If you like this tool, **consider starring the repository ✨** on GitHub! It
-really encourages me to build more open-source tools like this. :D
+I wanted a simple solution to calculate how much time I could save by using speed multipliers during lectures.
+But, I did not want to hop onto the browser or even open up the calculator, or even do mind-maths for this.
+Thus, I tried making a simple solution which lives in the terminal.
 
 ## Usage
-
-> [!NOTE]
-> For more information on all available commands, type `trimsec --help`.
-
-The tool provides two core functionalities: calculating saved time with a speed
-multiplier and managing your time bank data. This guide explains how to use both
-features in a seamless flow.
 
 To calculate saved time, you run the trim command as follows:
 
@@ -54,7 +34,7 @@ $ trimsec trim <duration> <speed>
 
 Here, `<duration>` represents the length of the video you are watching (using
 any combination of `<days>d`, `<hours>h`, `<minutes>m`, and `<seconds>s`), and
-`<speed>` is the speed multiplier—a floating-point number that can optionally be
+`<speed>` is the speed multiplier; a float that can optionally be
 suffixed with an `x` (for example, both `1.5x` and `1.5` are valid). For
 instance, if you are watching a 1-hour video at 2x speed, execute:
 
@@ -115,36 +95,24 @@ $ trimsec bank path
 
 ### cargo
 
-If you have [Rust](https://rust-lang.org/) and [Cargo](https://crates.io/)
-installed, you can install `trimsec` by running the following command:
-
 ```bash
 $ cargo install trimsec
 ```
 
 ### mise
 
-To install `trimsec` as a tool using [mise](https://github.com/jdx/mise), use
-the following command:
-
 ```bash
 $ mise use -g cargo:trimsec
 ```
 
-> [!NOTE]
-> These methods are subject to change as the library is in stable development.
-
----
-
-### Manual installation
+## Manual Installation
 
 Standalone binaries for this project are all available at the [GitHub
 Releases](https://github.com/hitblast/trimsec/releases) page. The binaries are
 produced by automated GitHub Actions workflows on three distinct runners and
 should run on all their respective platforms.
 
-In case, however, if you prefer building standalone binaries native to your
-machine, here's a few commands to help you:
+For manual building, proceed with the following:
 
 ```bash
 # Clone the repository.
@@ -155,26 +123,8 @@ $ cd trimsec
 $ cargo build --release
 ```
 
-Either you download the binary from the mentioned page above or manually
-compile, you will have to move it to somewhere which can be directly pointed at
-using the `$PATH` variable. Here's an example:
-
-```bash
-$ mv target/release/trimsec ~/.local/bin/
-
-# (Optional) This code appends the binary path to the PATH environment variable.
-$ echo 'export PATH="$HOME/.local/bin:$PATH"' >> ~/.bashrc
-$ source ~/.bashrc
-
-# Now you can run the binary from anywhere.
-$ trimsec 1h20m 1.75x
-```
-
 ## Contributors
 
-<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
-<!-- prettier-ignore-start -->
-<!-- markdownlint-disable -->
 <table>
   <tbody>
     <tr>
@@ -182,15 +132,6 @@ $ trimsec 1h20m 1.75x
     </tr>
   </tbody>
 </table>
-
-<!-- markdownlint-restore -->
-<!-- prettier-ignore-end -->
-
-<!-- ALL-CONTRIBUTORS-LIST:END -->
-
-## Changelog
-
-For release notes and changelog, please refer to the [CHANGELOG.md](https://github.com/hitblast/trimsec/blob/main/CHANGELOG.md) file.
 
 ## License
