+++ src/main.rs
@@ -1,5 +1,4 @@
-// Imports.
-use clap::{Parser, Subcommand};
+use clap::Parser;
 use colored::*;
 use std::{fs, process};
 
@@ -8,52 +7,10 @@ use trimsec::Config;
 mod time_bank;
 use time_bank::TimeBank;
 
-/// trimsec - Strategic (& fast) content consumption planner.
-#[derive(Parser)]
-#[command(author, version, about, long_about = None)]
-struct Cli {
-    #[command(subcommand)]
-    command: Commands,
-}
+mod args;
+use args::{BankCommands, Cli, Commands};
 
-#[derive(Subcommand)]
-enum Commands {
-    /// Calculate saved time using a multiplier over a given duration.
-    Trim {
-        /// The duration of the content (e.g. "1h30m" or "1h+30m").
-        duration: String,
-        /// The speed multiplier (e.g. "1.5x").
-        multiplier: String,
-        /// Only show new duration.
-        #[arg(short, long)]
-        duration_only: bool,
-        /// Only show saved time.
-        #[arg(short, long)]
-        time_saved_only: bool,
-        /// Use seconds as the time unit.
-        #[arg(short, long)]
-        seconds: bool,
-        /// Show emojis in the output.
-        #[arg(short, long)]
-        emoji: bool,
-    },
-    /// Manage or view your time bank data.
-    Bank {
-        #[command(subcommand)]
-        bank_command: Option<BankCommands>,
-    },
-}
-
-#[derive(Subcommand)]
-enum BankCommands {
-    /// Show the current time bank details.
-    Show,
-    /// Reset (clear) the time bank.
-    Reset,
-    /// Return the absolute path to the bank file.
-    Path,
-}
-
+// Helper functions for printing.
 fn print_error<T: std::fmt::Display>(msg: T) {
     eprintln!("{} {}", "[ERROR]".red(), msg);
 }
@@ -62,6 +19,7 @@ fn print_warning<T: std::fmt::Display>(msg: T) {
     eprintln!("{} {}", "[WARNING]".yellow(), msg);
 }
 
+// Main runner entrypoint.
 fn main() {
     let args = Cli::parse();
 
@@ -74,14 +32,12 @@ fn main() {
             seconds,
             emoji,
         } => {
-            // Create the Config – in case of errors exit.
             let config = Config::new(&duration, &multiplier).unwrap_or_else(|err| {
                 print_error(err);
                 process::exit(1);
             });
 
-            // Calculate using trimsec logic.
-            let result = trimsec::run(config);
+            let result = trimsec::trim(config);
             match result {
                 Ok((new_duration, time_saved, splits)) => {
                     // Display new duration.
@@ -173,56 +129,50 @@ fn main() {
             }
         }
 
-        Commands::Bank { bank_command } => {
-            match bank_command {
-                Some(BankCommands::Show) | None => {
-                    // By default, show time bank details.
-                    match TimeBank::load() {
-                        Ok(bank) => {
-                            if bank.entries.is_empty() {
-                                println!("Time bank is empty.");
-                            } else {
-                                println!("Time Bank Details:");
-                                for entry in &bank.entries {
-                                    println!(
-                                        "  {}: {}",
-                                        entry.date,
-                                        trimsec::parse_time(entry.saved_time)
-                                    );
-                                }
-                                println!(
-                                    "\nTotal saved time: {}",
-                                    trimsec::parse_time(bank.total_saved())
-                                );
-                            }
-                        }
-                        Err(e) => {
-                            print_error(format!("Could not load time bank: {}", e));
-                        }
-                    }
-                }
-                Some(BankCommands::Reset) => {
-                    // Reset the bank: overwrite with an empty bank structure.
-                    let bank = TimeBank { entries: vec![] };
-                    if let Err(e) = bank.save() {
-                        print_error(format!("Could not reset time bank: {}", e));
-                        process::exit(1);
+        Commands::Bank { bank_command } => match bank_command {
+            Some(BankCommands::Show) | None => match TimeBank::load() {
+                Ok(bank) => {
+                    if bank.entries.is_empty() {
+                        println!("Time bank is empty.");
                     } else {
-                        println!("Time bank has been reset.");
+                        println!("Time Bank Details:");
+                        for entry in &bank.entries {
+                            println!(
+                                "  {}: {}",
+                                entry.date,
+                                trimsec::parse_time(entry.saved_time)
+                            );
+                        }
+                        println!(
+                            "\nTotal saved time: {}",
+                            trimsec::parse_time(bank.total_saved())
+                        );
                     }
                 }
-                Some(BankCommands::Path) => match TimeBank::load() {
-                    Ok(_) => match fs::canonicalize(TimeBank::bank_file_path()) {
-                        Ok(path) => println!("{}", path.display()),
-                        Err(_) => {
-                            print_error("Could not get canonical path. Time bank was not initialized by trimsec.");
-                        }
-                    },
-                    Err(e) => {
-                        print_error(format!("Could not load time bank: {}", e));
+                Err(e) => {
+                    print_error(format!("Could not load time bank: {}", e));
+                }
+            },
+            Some(BankCommands::Reset) => {
+                let bank = TimeBank { entries: vec![] };
+                if let Err(e) = bank.save() {
+                    print_error(format!("Could not reset time bank: {}", e));
+                    process::exit(1);
+                } else {
+                    println!("Time bank has been reset.");
+                }
+            }
+            Some(BankCommands::Path) => match TimeBank::load() {
+                Ok(_) => match fs::canonicalize(TimeBank::bank_file_path()) {
+                    Ok(path) => println!("{}", path.display()),
+                    Err(_) => {
+                        print_error("Could not get canonical path. Time bank was not initialized by trimsec.");
                     }
                 },
-            }
-        }
+                Err(e) => {
+                    print_error(format!("Could not load time bank: {}", e));
+                }
+            },
+        },
     }
 }
